--- arco-design-pro-next/src/access.ts
@@ -1,7 +1,0 @@
-export const isAdmin = (roles: string[]) => {
-  return roles.includes('ROLE_ADMIN');
-}
-
-export const isNormalUser = (roles: string[]) => {
-  return roles.includes('ROLE_USER');
-}
